refactor: arrange package
