+++ nextstep-jdbc/src/main/java/nextstep/jdbc/db/BasicDataSourceFactory.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.db;
 
 import org.apache.commons.dbcp2.BasicDataSource;
