+++ nextstep-jdbc/src/main/java/nextstep/jdbc/template/PreparedStatementHandler.java
@@ -1,9 +1,9 @@
-package nextstep.jdbc;
+package nextstep.jdbc.template;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 @FunctionalInterface
-public interface Handler<T> {
+public interface PreparedStatementHandler<T> {
     T handle(PreparedStatement pstmt) throws SQLException;
 }
