+++ nextstep-jdbc/src/main/java/nextstep/jdbc/template/PreparedStatementSetter.java
@@ -1,8 +1,9 @@
-package nextstep.jdbc;
+package nextstep.jdbc.template;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
+@Deprecated
 public interface PreparedStatementSetter {
     void setValues(PreparedStatement pstmt) throws SQLException;
 }
